+++ server.js
@@ -26,10 +26,12 @@ let wall = new Wall();
 let castle = new Building();
 let warehouse = new Building();
 let windmill = new Building();
+let barracks = new Building();
 
 castle.define('castle', [ {name: 'iron', amount: 1}, {name: 'wood', amount: 1}, {name: 'clay', amount: 1}, {name: 'grain', amount: 1} ]);
 warehouse.define('warehouse', [ {name: 'iron', amount: 1}, {name: 'wood', amount: 2}, {name: 'clay', amount: 1}, {name: 'grain', amount: 2}, ]);
 windmill.define('windmill', [ {name: 'iron', amount: 2}, {name: 'wood', amount: 3}, {name: 'clay', amount: 3}, {name: 'grain', amount: 1}, ]);
+barracks.define('windmill', [ {name: 'iron', amount: 0}, {name: 'wood', amount: 3}, {name: 'clay', amount: 3}, {name: 'grain', amount: 1}, ]);
 
 // websockets
 server.on('connection', ws => {
@@ -43,8 +45,9 @@ server.on('connection', ws => {
 tree.addBuilding('castle', castle);
 tree.addBuilding('warehouse', warehouse, 1, 'castle', 1);
 tree.addBuilding('windmill', windmill, 1, 'castle', 1);
-tree.addBuilding('castle', castle, 1, 'windmill', 1);
-tree.addBuilding('castle', castle, 1, 'warehouse', 1);
+tree.addBuilding('castle', castle, 2, 'windmill', 1);
+tree.addBuilding('castle', castle, 2, 'warehouse', 1);
+tree.addBuilding('barracks', barracks, 1, 'castle', 2);
 
 //{ name: 'castle', level: 1 },
 //{ name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },