+++ src/wall/wall.test.js
@@ -1,7 +1,7 @@
 const Wall = require('./wall')
-const wall = new Wall();
 
 test(`list of available actions`, () => {
+    const wall = new Wall();
     wall.treeBuilding([
         { name: 'castle', level: 1 },
         { name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },
@@ -16,6 +16,7 @@ test(`list of available actions`, () => {
 })
 
 test(`deny building that is not present in the tree`, () => {
+    const wall = new Wall();
     wall.treeBuilding([
         { name: 'castle', level: 1 },
         { name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },
@@ -29,6 +30,7 @@ test(`deny building that is not present in the tree`, () => {
 test(`allow next level of an already built building`, () => { })
 
 test(`list of available actions`, () => {
+    const wall = new Wall();
     wall.treeBuilding([
         { name: 'castle', level: 1 },
         { name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },
@@ -44,6 +46,7 @@ test(`list of available actions`, () => {
 })
 
 test(`list of available actions`, () => {
+    const wall = new Wall();
     wall.treeBuilding([
         { name: 'castle', level: 1 },
         { name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },
@@ -54,4 +57,18 @@ test(`list of available actions`, () => {
     wall.addToQueue({ name: 'castle', level: 1 });
     wall.addToQueue({ name: 'warehouse', level: 1 });
     expect(wall.canBuild('warehouse', 1)).toEqual(false)
+})
+
+test(`extracrt next level of a given building`, () => {
+    const wall = new Wall();
+    wall.treeBuilding([
+        { name: 'castle', level: 1 },
+        { name: 'windmill', level: 1, required: { name: 'castle', level: 1 } },
+        { name: 'warehouse', level: 1, required: { name: 'castle', level: 1 } },
+        { name: 'castle', level: 2, required: { name: 'windmill', level: 1 } },
+        { name: 'castle', level: 2, required: { name: 'warehouse', level: 1 } },
+    ]);
+    wall.addToQueue({ name: 'castle', level: 1 });
+    wall.addToQueue({ name: 'castle', level: 2 });
+    expect(wall.extractNextLevelOf('castle')).toEqual(3);
 })