+++ CHANGELOG.md
@@ -18,3 +18,4 @@
  - remove client from this repository
  - server send tree data to the client
  - provide distinct list of buildings inserted in the tree
+ - detect next level of a certain building